+++ index.js
@@ -54,9 +54,39 @@ let base = {
         "editorGutter.addedBackground": "#608b4e",
         "editorGutter.modifiedBackground": "#d7ba7d",
         "editorGutter.deletedBackground": "#d16969",
+
         "editorOverviewRuler.addedForeground": "#608b4e",
         "editorOverviewRuler.modifiedForeground": "#d7ba7d",
         "editorOverviewRuler.deletedForeground": "#d16969",
+
+        // "editorOverviewRuler.border": "", // Color of the overview ruler border.
+        // "editorOverviewRuler.findMatchForeground": "", // Overview ruler marker color for find matches. The color must not be opaque to not hide underlying decorations.
+        // "editorOverviewRuler.rangeHighlightForeground": "", //Overview ruler marker color for highlighted ranges, like by the Quick Open, Symbol in File and Find features. The color must not be opaque to not hide underlying decorations.
+        // "editorOverviewRuler.selectionHighlightForeground": "", //Overview ruler marker color for selection highlights. The color must not be opaque to not hide underlying decorations.
+        // "editorOverviewRuler.wordHighlightForeground": "", //Overview ruler marker color for symbol highlights. The color must not be opaque to not hide underlying decorations.
+        // "editorOverviewRuler.wordHighlightStrongForeground": "", // Overview ruler marker color for write-access symbol highlights. The color must not be opaque to not hide underlying decorations.
+        // "editorOverviewRuler.modifiedForeground": "", // Overview ruler marker color for modified content.
+        // "editorOverviewRuler.addedForeground": "", // Overview ruler marker color for added content.
+        // "editorOverviewRuler.deletedForeground": "", // Overview ruler marker color for deleted content.
+        // "editorOverviewRuler.errorForeground": "", // Overview ruler marker color for errors.
+        // "editorOverviewRuler.warningForeground": "", // Overview ruler marker color for warnings.
+        // "editorOverviewRuler.infoForeground": "", // Overview ruler marker color for infos.
+
+        // "editorError.foreground": "#d16969",
+        // "editorError.border": "#d16969",
+        // "editorWarning.foreground": "#d7ba7d",
+        // "editorWarning.border": "#d7ba7d",
+        // "editorInfo.foreground": "#608b4e",
+        // "editorInfo.border": "#608b4e",
+
+        // "notificationCenter.border": "", // Notification Center border color.
+        // "notificationCenterHeader.foreground": "", // Notification Center header foreground color.
+        // "notificationCenterHeader.background": "", // Notification Center header background color.
+        // "notificationToast.border": "", // Notification toast border color.
+        // "notifications.foreground": "", // Notification foreground color.
+        // "notifications.background": "", // Notification background color.
+        // "notifications.border": "", // Notification border color separating from other notifications in the Notification Center.
+        // "notificationLink.foreground": "", // Notification links foreground color.
 
         "gitDecoration.modifiedResourceForeground": "#608b4e",
         "gitDecoration.deletedResourceForeground": "#d7ba7d",