+++ src/partials/html.js
@@ -3,18 +3,18 @@ module.exports = [
         "scope": [
             "punctuation.definition.tag",
         ],
-        "settings": { "foreground": "#808080" }
+        "settings": { "foreground": "#808080", },
     },
     {
         "scope": "entity.other.attribute-name",
-        "settings": { "foreground": "#9cdcfe" }
+        "settings": { "foreground": "#9cdcfe", },
     },
     {
         "scope": [
             "punctuation.definition.string.begin.html",
             "punctuation.definition.string.end.html",
         ],
-        "settings": { "foreground": "#ce9178" }
+        "settings": { "foreground": "#ce9178", },
     },
     {
         "scope": [
@@ -30,7 +30,7 @@ module.exports = [
             "meta.tag.template.value.twig",
             "punctuation.section.tag.twig",
         ],
-        "settings": { "foreground": "#569cd6" }
+        "settings": { "foreground": "#569cd6", },
     },
     {
         "scope": [
@@ -49,7 +49,7 @@ module.exports = [
             "text.html.jinja punctuation.definition.block.js",
             "text.html.jinja punctuation.separator.comma.js",
         ],
-        "settings": { "foreground": "#d4d4d4" }
+        "settings": { "foreground": "#d4d4d4", },
     },
     // {
     //     "name": "Template Tags",
@@ -64,12 +64,12 @@ module.exports = [
     // },
     {
         "scope": "constant.character.entity.html",
-        "settings": { "foreground": "#dcdcaa" }
+        "settings": { "foreground": "#dcdcaa", },
     },
     {
         "name": "don't highlight ejs and erb as errors",
         "scope": "invalid.illegal.bad-angle-bracket.html",
-        "settings": { "foreground": "#d4d4d4" }
+        "settings": { "foreground": "#d4d4d4", },
     },
     {
         "name": "Mustaches",
@@ -79,7 +79,7 @@ module.exports = [
             "punctuation.definition.generic.begin.html",
             "punctuation.definition.generic.end.html",
         ],
-        "settings": { "foreground": "#4ec9b0" }
+        "settings": { "foreground": "#4ec9b0", },
     },
     // =========================================================================
     // Jade/Pug
@@ -89,7 +89,7 @@ module.exports = [
         "scope": [
             "string.interpolated.jade",
         ],
-        "settings": { "foreground": "#4ec9b0" }
+        "settings": { "foreground": "#4ec9b0", },
     },
     {
         "scope": [
@@ -98,7 +98,7 @@ module.exports = [
             "storage.type.function.jade",
             "storage.type.import.include.jade",
         ],
-        "settings": { "foreground": "#c586c0" }
+        "settings": { "foreground": "#c586c0", },
     },
     // =========================================================================
     // Handlebars
@@ -116,19 +116,19 @@ module.exports = [
             "meta.function.block.start.handlebars support.constant.handlebars",
             "meta.function.block.end.handlebars support.constant.handlebars",
         ],
-        "settings": { "foreground": "#c586c0" }
+        "settings": { "foreground": "#c586c0", },
     },
     {
         "scope": [
             "support.constant.handlebars",
         ],
-        "settings": { "foreground": "#4ec9b0" }
+        "settings": { "foreground": "#4ec9b0", },
     },
     {
         "scope": [
             "variable.parameter.handlebars",
         ],
-        "settings": { "foreground": "#9cdcfe" }
+        "settings": { "foreground": "#9cdcfe", },
     },
     // =========================================================================
     // ejs
@@ -139,7 +139,7 @@ module.exports = [
             // "text.html.ejs keyword.begin.tag.ejs punctuation.accessor.js",
             // "text.html.ejs keyword.begin.tag.ejs keyword.operator.arithmetic.js",
         ],
-        "settings": { "foreground": "#569cd6" }
+        "settings": { "foreground": "#569cd6", },
     },
     {
         "scope": [
@@ -149,29 +149,29 @@ module.exports = [
             "text.html.ejs punctuation.definition.parameters.begin.js",
             "text.html.ejs punctuation.definition.parameters.end.js",
         ],
-        "settings": { "foreground": "#d4d4d4" }
+        "settings": { "foreground": "#d4d4d4", },
     },
     {
         "scope": [
             "text.html.ejs keyword.operator.expression.typeof.js",
             "text.html.ejs keyword.operator.expression.instanceof.js",
         ],
-        "settings": { "foreground": "#c586c0" }
+        "settings": { "foreground": "#c586c0", },
     },
     // =========================================================================
     // Vue
     // =========================================================================
     {
         "scope": [
-            "text.html.vue-html meta.brace.round.js"
+            "text.html.vue-html meta.brace.round.js",
         ],
-        "settings": { "foreground": "#d4d4d4" }
+        "settings": { "foreground": "#d4d4d4", },
     },
     {
         "scope": [
-            "text.html.vue-html meta.tag.other.html"
+            "text.html.vue-html meta.tag.other.html",
         ],
-        "settings": { "foreground": "#808080" }
+        "settings": { "foreground": "#808080", },
     },
     {
         // punctuation.definition.string.begin.html
@@ -183,34 +183,34 @@ module.exports = [
             "text.html.vue-html source.directive.vue punctuation.definition.string.begin.html",
             "text.html.vue-html source.directive.vue punctuation.definition.string.end.html",
         ],
-        "settings": { "foreground": "#4ec9b0" }
+        "settings": { "foreground": "#4ec9b0", },
     },
     // =========================================================================
     // Jinja2
     // =========================================================================
     {
         "scope": [
-            "entity.other.jinja.delimiter.tag"
+            "entity.other.jinja.delimiter.tag",
         ],
-        "settings": { "foreground": "#606060" }
+        "settings": { "foreground": "#606060", },
     },
     {
         "scope": [
-            "entity.other.jinja.delimiter.variable"
+            "entity.other.jinja.delimiter.variable",
         ],
-        "settings": { "foreground": "#4ec9b0" }
+        "settings": { "foreground": "#4ec9b0", },
     },
     {
         "scope": [
-            "variable.other.jinja"
+            "variable.other.jinja",
             // "variable.other.jinja.block"
         ],
-        "settings": { "foreground": "#9cdcfe" }
+        "settings": { "foreground": "#9cdcfe", },
     },
     {
         "scope": [
             "variable.language.jinja",
         ],
-        "settings": { "foreground": "#569cd6" }
+        "settings": { "foreground": "#569cd6", },
     },
-]
+];